% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyLG_functions.R
\docType{package}
\name{easyLG}
\alias{easyLG}
\alias{easyLG-package}
\alias{easyLG-package}
\title{easyLG : Landscape genomics made easy!}
\description{
This R package implement a Shiny app for Landscape Genomic analysis.
The raw app could be find at \url{https://github.com/aariani/LGapp}.
}
\details{
easyLG : Landscape Genomics made easy

For a comprehensive tutorial see \url{https://github.com/aariani/easyLG/wiki}
}
