% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyLG_functions.R
\name{easyLG_app}
\alias{easyLG_app}
\title{R shiny app for landscape genomic}
\usage{
easyLG_app()
}
\description{
Launch the easyLG shiny application

For a comprehensive tutorial see \url{https://github.com/aariani/easyLG/wiki}
}
