% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyLG_functions.R
\name{easyLG_check}
\alias{easyLG_check}
\title{Check and install dependencies}
\usage{
easyLG_check()
}
\description{
Install the packages necessary for using the easyLG_app function. Before running made sure to install all the dependencies outside R as described in \url{ADD URL}
}
